feat: added a soft delete features, fixed some bugs and created more models
