+++ src/routes/user.routes.js
@@ -12,6 +12,6 @@ router.post('/users', createAUser);
 router.get('/users', fetchAllUser);
 router.delete('/users/:id', DeleteAUser);
 router.get('/users/:id', fetchAUser);
-router.put('/nodemusers/:id', editAUser);
+router.put('/users/:id', editAUser);
 
 module.exports = router;
